[REFACTOR] Magazine 구조체 따로 뺴주기 #6
